sneaker_webserver/pyproject.toml
[tool.poetry]
name = "sneaker_webserver"
version = "0.1.0"
description = ""
authors = ["Daniel Lietz <ge96jek@mytum.de>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.4"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.2.0"
Pillow = "^10.0.0"
requests = "^2.31.0"
httpx = {extras = ["http2"], version = "^0.24.1"}
nested-lookup = "^0.2.25"
parsel = "^1.8.1"
asgiref = "^3.7.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
